Struct startuppong::Account

pub struct Account {
    // some fields omitted
}

An account is necessary to make requests against the API.

This struct holds your account ID and access key. It is a required argument to all of the API methods.

Methods

impl Account

fn new(id: String, key: String) -> Account

Create a new Account

fn from_env() -> Result<Account, VarError>

Try and create an account from environment variables

The environment variables needed are STARTUPPONG_ACCOUNT_ID and STARTUPPONG_ACCESS_KEY.

fn id(&self) -> &str

Return a ref to the api_account_id

fn key(&self) -> &str

Return a ref to the api_access_key