Function startuppong::add_match_with_names

pub fn add_match_with_names(account: &Account, winner: &str, loser: &str) -> Result<(), ApiError>

Add a match using the given names.

It is possible for name lookup to fail. If that happens, the Result will unwrap to a startuppong::ApiError::PlayerNotFound(String) where the String is the first name not resolved. Names are matched in a case-sensitive fashion. The first result with a valid sub string match is accepted.